# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:8083
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /servicio/cometra:
    x-swagger-router-controller: consultas
    get:
      operationId: getQuerieCometra
      description: consulta los marcajes por rango de fechas
      parameters:
        - name: rango
          description: Rango de fechas
          in: body
          required: true
          schema:
            $ref: "#/definitions/Rango"
      responses:
        "200":
          description: Success
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /servicio/sepsa:
    x-swagger-router-controller: consultas
    get:
      operationId: getQuerieSepsa
      description: consulta los marcajes por rango de fechas
      parameters:
        - name: rango
          description: Rango de fechas
          in: body
          required: true
          schema:
            $ref: "#/definitions/Rango"
      responses:
        "200":
          description: Success
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /servicio/seguritec:
    x-swagger-router-controller: consultas
    get:
      operationId: getQuerieSeg
      description: consulta los marcajes por rango de fechas
      parameters:
        - name: rango
          description: Rango de fechas
          in: body
          required: true
          schema:
            $ref: "#/definitions/Rango"
      responses:
        "200":
          description: Success
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Rango:
    properties:
      fechaInicio:
        type: string
      fechaFin:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string